Make $manage a boolean parameter and remove document root management
NOTE: this is an incompatible change and may cause issues in modules that
previously depended on the document root being created by this module. However,
as the document root is specific to each virtual host, it did not really belong
into this module. The $manage parameter will no longer accept 'yes' as its
value. This change, however, will at most break Puppet runs, because the
booleanness of $manage is validated before applying the catalog.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>